refactor: update `getAll()` return type to Flow in RecentWatchedDao
