+++ data/src/main/java/com/london/data/local/database/dao/recent/watched/RecentWatchedDao.kt
@@ -1,0 +1,11 @@
+@file:KoverIgnore
+
+package com.london.data.local.database.dao.recent.watched
+
+import com.london.domain.KoverIgnore
+import kotlinx.coroutines.flow.Flow
+
+interface RecentWatchedDao<T> {
+    fun getAll(): Flow<List<T>>
+    suspend fun insert(item: T)
+}
