+++ data/src/main/java/com/london/data/local/database/dao/recent/watched/movie/RecentWatchedMoviesDao.kt
@@ -1,20 +1,21 @@
 @file:KoverIgnore
 
-package com.london.data.local.database.dao.recent.whatched.movie
+package com.london.data.local.database.dao.recent.watched.movie
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.london.data.local.database.dao.recent.whatched.RecentWatchedDao
+import com.london.data.local.database.dao.recent.watched.RecentWatchedDao
 import com.london.data.local.model.recent.watched.RecentWatchedMovieLocal
 import com.london.domain.KoverIgnore
+import kotlinx.coroutines.flow.Flow
 
 @Dao
-interface RecentWatchedMoviesDao: RecentWatchedDao<RecentWatchedMovieLocal>{
+interface RecentWatchedMoviesDao : RecentWatchedDao<RecentWatchedMovieLocal> {
 
     @Query("SELECT * FROM recent_watched_movie_table")
-    override suspend fun getAll(): List<RecentWatchedMovieLocal>
+    override fun getAll(): Flow<List<RecentWatchedMovieLocal>>
 
     @Insert(onConflict = OnConflictStrategy.Companion.REPLACE)
     override suspend fun insert(item: RecentWatchedMovieLocal)
