+++ data/src/main/java/com/london/data/local/database/dao/recent/watched/tvshow/RecentWatchedTvShowsDao.kt
@@ -1,17 +1,18 @@
-package com.london.data.local.database.dao.recent.whatched.tvshow
+package com.london.data.local.database.dao.recent.watched.tvshow
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.london.data.local.database.dao.recent.whatched.RecentWatchedDao
+import com.london.data.local.database.dao.recent.watched.RecentWatchedDao
 import com.london.data.local.model.recent.watched.RecentWatchedTvShowLocal
+import kotlinx.coroutines.flow.Flow
 
 @Dao
-interface RecentWatchedTvShowsDao: RecentWatchedDao<RecentWatchedTvShowLocal>{
+interface RecentWatchedTvShowsDao : RecentWatchedDao<RecentWatchedTvShowLocal> {
 
     @Query("SELECT * FROM recent_watched_tv_show_table")
-    override suspend fun getAll(): List<RecentWatchedTvShowLocal>
+    override fun getAll(): Flow<List<RecentWatchedTvShowLocal>>
 
     @Insert(onConflict = OnConflictStrategy.Companion.REPLACE)
     override suspend fun insert(item: RecentWatchedTvShowLocal)
